requirements.txt
azure-common==1.1.28
azure-core==1.31.0
azure-cosmos==4.7.0
azure-identity==1.18.0
azure-mgmt-applicationinsights==4.0.0
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.1.1
fastapi==0.115.0
uvicorn==0.31.0